add util method
